+++ src/main/scala/com/guizmaii/distances/GeoCache.scala
@@ -1,4 +1,4 @@
-package com.guizmaii.distances.implementations.cache
+package com.guizmaii.distances
 
 import monix.eval.Task
 
@@ -18,8 +18,6 @@ abstract class GeoCache[E <: Serializable: ClassTag] {
 
   implicit val innerCache: Cache[E]
 
-  def getOrTask(keyParts: Any*)(orElse: => Task[E]): Task[E] = cachingF(keyParts)(ttl = Some(expiration))(orElse)
-
-  def setT(keyParts: Any*)(value: E): Task[Any] = put(keyParts)(value = value, ttl = Some(expiration))
+  final def getOrTask(keyParts: Any*)(orElse: => Task[E]): Task[E] = cachingF(keyParts)(ttl = Some(expiration))(orElse)
 
 }