+++ src/main/scala/com/guizmaii/distances/Geocoder.scala
@@ -1,11 +1,20 @@
 package com.guizmaii.distances
 
-import com.guizmaii.distances.Types.{LatLong, PostalCode}
-import com.guizmaii.distances.utils.WithCache
+import com.guizmaii.distances.Types.{Address, LatLong, PostalCode}
 import monix.eval.Task
 import monix.execution.CancelableFuture
 
-trait Geocoder extends WithCache[LatLong] {
-  def geocodeT(postalCode: PostalCode): Task[LatLong]
-  def geocode(postalCode: PostalCode): CancelableFuture[LatLong]
+trait Geocoder {
+
+  def geocodePostalCodeT(postalCode: PostalCode)(implicit cache: GeoCache[LatLong]): Task[LatLong]
+  def geocodePostalCode(postalCode: PostalCode)(implicit cache: GeoCache[LatLong]): CancelableFuture[LatLong]
+
+  /**
+    * Doc about "non ambigue addresses": https://developers.google.com/maps/documentation/geocoding/best-practices#complete-address
+    *
+    * @param address
+    * @return
+    */
+  def geocodeNonAmbigueAddressT(address: Address)(implicit cache: GeoCache[LatLong]): Task[LatLong]
+  def geocodeNonAmbigueAddress(address: Address)(implicit cache: GeoCache[LatLong]): CancelableFuture[LatLong]
 }